+++ src/index.ts
@@ -1,5 +1,8 @@
 export { TinkoffInvestNodeSDK } from './bootstrap/tinkoff-invest-node-sdk';
-export type { TinkoffInvestOptions } from './application/dto/tinkoff-invest-options';
+export type {
+  TinkoffInvestOptions,
+  TinkoffInvestTlsOptions
+} from './application/dto/tinkoff-invest-options';
 export {
   isSdkError,
   SdkError,